# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Linguascreen language learning app",
        "title": "Linguascreen API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/explain": {
            "post": {
                "description": "Explain and translate selected text using LLM and NMT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Explain selected text",
                "parameters": [
                    {
                        "description": "Explain request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.PostExplainRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.PostExplainResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ocr": {
            "post": {
                "description": "Extract OCR text from base64 image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Extract text from image",
                "parameters": [
                    {
                        "description": "OCR request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.PostOcrRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.PostOcrResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/quiz/{user_id}": {
            "get": {
                "description": "Generate a quiz based on weakest memories for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get quiz for user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.GetQuizResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Baseline": {
            "type": "object",
            "properties": {
                "x0": {
                    "type": "number"
                },
                "x1": {
                    "type": "number"
                },
                "y0": {
                    "type": "number"
                },
                "y1": {
                    "type": "number"
                }
            }
        },
        "models.Bbox": {
            "type": "object",
            "properties": {
                "x0": {
                    "type": "number"
                },
                "x1": {
                    "type": "number"
                },
                "y0": {
                    "type": "number"
                },
                "y1": {
                    "type": "number"
                }
            }
        },
        "models.Block": {
            "type": "object",
            "properties": {
                "bbox": {
                    "$ref": "#/definitions/models.Bbox"
                },
                "blocktype": {
                    "type": "integer"
                },
                "confidence": {
                    "type": "number"
                },
                "paragraphs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Paragraph"
                    }
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.Choice": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.ExplainResult": {
            "type": "object",
            "properties": {
                "grammar_explanations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MemoryInput"
                    }
                },
                "phrase_explanations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MemoryInput"
                    }
                },
                "word_explanations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MemoryInput"
                    }
                }
            }
        },
        "models.Line": {
            "type": "object",
            "properties": {
                "baseline": {
                    "$ref": "#/definitions/models.Baseline"
                },
                "bbox": {
                    "$ref": "#/definitions/models.Bbox"
                },
                "confidence": {
                    "type": "number"
                },
                "rowAttributes": {
                    "$ref": "#/definitions/models.RowAttributes"
                },
                "text": {
                    "type": "string"
                },
                "words": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Word"
                    }
                }
            }
        },
        "models.MemoryInput": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "example_sentence": {
                    "type": "string"
                },
                "explanation": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "translation": {
                    "type": "string"
                }
            }
        },
        "models.Page": {
            "type": "object",
            "properties": {
                "blocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Block"
                    }
                },
                "box": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                },
                "debug": {
                    "type": "string"
                },
                "hocr": {
                    "type": "string"
                },
                "imageBinary": {
                    "type": "string"
                },
                "imageColor": {
                    "type": "string"
                },
                "imageGrey": {
                    "type": "string"
                },
                "oem": {
                    "type": "string"
                },
                "osd": {
                    "type": "string"
                },
                "pdf": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "psm": {
                    "type": "string"
                },
                "rotateRadians": {
                    "type": "number"
                },
                "sd": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "tsv": {
                    "type": "string"
                },
                "unlv": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.Paragraph": {
            "type": "object",
            "properties": {
                "bbox": {
                    "$ref": "#/definitions/models.Bbox"
                },
                "confidence": {
                    "type": "number"
                },
                "is_ltr": {
                    "type": "integer"
                },
                "lines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Line"
                    }
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.Quiz": {
            "type": "object",
            "properties": {
                "questionCount": {
                    "type": "integer"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.QuizQuestion"
                    }
                }
            }
        },
        "models.QuizQuestion": {
            "type": "object",
            "properties": {
                "answerID": {
                    "type": "integer"
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "memoryID": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "models.RowAttributes": {
            "type": "object",
            "properties": {
                "ascenders": {
                    "type": "number"
                },
                "descenders": {
                    "type": "number"
                },
                "rowHeight": {
                    "type": "number"
                }
            }
        },
        "models.Symbol": {
            "type": "object",
            "properties": {
                "bbox": {
                    "$ref": "#/definitions/models.Bbox"
                },
                "confidence": {
                    "type": "number"
                },
                "is_dropcap": {
                    "type": "integer"
                },
                "is_subscript": {
                    "type": "integer"
                },
                "is_superscript": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.Word": {
            "type": "object",
            "properties": {
                "bbox": {
                    "$ref": "#/definitions/models.Bbox"
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Choice"
                    }
                },
                "confidence": {
                    "type": "number"
                },
                "font_name": {
                    "type": "string"
                },
                "symbols": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Symbol"
                    }
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "routes.GetQuizResponse": {
            "type": "object",
            "properties": {
                "quiz": {
                    "$ref": "#/definitions/models.Quiz"
                }
            }
        },
        "routes.PostExplainRequest": {
            "type": "object",
            "required": [
                "selected_text"
            ],
            "properties": {
                "selected_text": {
                    "type": "string",
                    "example": "Hello world"
                },
                "target_lang": {
                    "type": "string",
                    "example": "es"
                },
                "user_lang": {
                    "type": "string",
                    "example": "en"
                }
            }
        },
        "routes.PostExplainResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "$ref": "#/definitions/models.ExplainResult"
                }
            }
        },
        "routes.PostOcrRequest": {
            "type": "object",
            "required": [
                "base64Image",
                "lang"
            ],
            "properties": {
                "base64Image": {
                    "type": "string",
                    "example": "base64string"
                },
                "confidenceThreshold": {
                    "type": "number",
                    "example": 0.8
                },
                "lang": {}
            }
        },
        "routes.PostOcrResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "$ref": "#/definitions/models.Page"
                },
                "text": {
                    "type": "string"
                }
            }
        }
    }
}